requirements.txt
feedparser==6.0.10
pyyaml==6.0.3
requests==2.31.0
